services/vector-memory-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
openai==1.3.0
faiss-cpu==1.7.4
numpy==1.24.3
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0

